import: Add system tagging support
OpenVPN 3 Linux v21 adds supports for configuration profile tags.  When
importing a configuration profile, the profile can be given a single
system tag.

This change checks if the configuration manager version is recent enough
and will add the needed tag value if it is supported.

A system tag is designed to be static per implementation, for
openvpn3-indicator, the tag 'ovpn3indc' was randomly chosen.

To search for configurations imported by OpenVPN 3 Indicator:

    $ openvpn3 configs-list --filter-tag system:ovpn3indc

Signed-off-by: David Sommerseth <davids@openvpn.net>